コンポーネントの切り出し
